server/requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-filter>=23.0,<24.0
djangorestframework-simplejwt>=5.2,<6.0
psycopg2-binary>=2.9,<3.0  # If using PostgreSQL
Pillow>=10.0,<11.0         # For image processing
celery>=5.3,<6.0           # For background tasks
redis>=5.0,<6.0            # For caching/task queue
python-decouple>=3.8,<4.0  # For environment variable management
django-cors-headers>=4.3,<5.0  # For CORS support
drf-spectacular>=0.27,<1.0  # For OpenAPI/Swagger documentation
cloudinary>=1.34,<2.0
